[game] Player information refactor - Separated Hand from Profile - Created section for the timer / profiles in match - Color indicators and placeholder for additional player metadata
