+++ client/src/components/GameBoard/PlayerPanel.css
@@ -1,0 +1,33 @@
+.player__profile {
+  border: solid 1px #ececec;
+  width: 300px;
+  height: 180px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+}
+
+.player__color {
+  width: 20px;
+  height: 20px;
+  margin: 8px;
+  border-radius: 10px;
+  border: solid 1px #ececec;
+}
+.white {
+  background-color: #fff;
+}
+.black {
+  background-color: #000;
+}
+
+.player__name {
+  margin: 4px;
+  font-size: 18px;
+  font-weight: 800;
+}
+.player__stats {
+  font-size: 14px;
+  font-style: italic;
+}
